requirements.txt
# 基础依赖
carla>=0.9.13
numpy>=1.19.0
torch>=1.7.0
opencv-python>=4.5.0
pygame>=2.0.0
wandb>=0.12.0
matplotlib>=3.3.0
pyyaml>=5.4.0
tensorboard>=2.4.0
tqdm>=4.60.0
psutil>=5.8.0

# 开发依赖
pytest>=6.0
pytest-cov>=2.0
black>=21.0
flake8>=3.9.0
mypy>=0.910
pre-commit>=2.15.0
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0 